requirements.txt
# Web Framework
Flask==3.0.0

# Computer Vision & Image Processing
opencv-python==4.9.0.80
fer==22.5.1
moviepy==1.0.3
Pillow==10.2.0
numpy==1.26.4

# Environment & Configuration
python-dotenv==1.0.1

# Music Integration
spotipy==2.24.0

# Database (Stable Supabase version)
supabase==1.0.3

# Data Models
dataclasses-json==0.6.4

# TensorFlow
tensorflow==2.17.0
